Icons CSS. `propTypes` on `ActionBar` and `Icon`. `ActionBar` in separate file.
