+++ src/xo-app/vm/index.js
@@ -1,140 +1,10 @@
 import _ from 'messages'
-import ActionBar from 'action-bar'
 import React, { Component } from 'react'
 import xo from 'xo'
 import { Row, Col } from 'grid'
 import { connectStore } from 'utils'
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs'
-
-// ===================================================================
-
-const vmActionBarByState = {
-  Running: ({ handlers, vm }) => (
-    <ActionBar
-      actions={[
-        {
-          icon: 'stop',
-          label: 'stopVmLabel',
-          handler: () => handlers.stopVm(vm.id)
-        },
-        {
-          icon: 'suspend',
-          label: 'suspendVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'reboot',
-          label: 'rebootVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'migrate',
-          label: 'migrateVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'force-reboot',
-          label: 'forceRebootVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'delete',
-          label: 'deleteVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'force-shutdown',
-          label: 'forceShutdownVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'snapshot',
-          label: 'snapshotVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'export',
-          label: 'exportVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'copy',
-          label: 'copyVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'console',
-          label: 'vmConsoleLabel',
-          handler: () => {}
-        }
-      ]}
-    />
-  ),
-  Halted: ({ handlers, vm }) => (
-    <ActionBar
-      actions={[
-        {
-          icon: 'start',
-          label: 'startVmLabel',
-          handler: () => handlers.startVm(vm.id)
-        },
-        {
-          icon: 'recovery-mode',
-          label: 'recoveryModeLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'delete',
-          label: 'deleteVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'clone',
-          label: 'cloneVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'create-template',
-          label: 'convertToTemplateLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'snapshot',
-          label: 'snapshotVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'export',
-          label: 'exportVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'copy',
-          label: 'copyVmLabel',
-          handler: () => {}
-        },
-        {
-          icon: 'console',
-          label: 'vmConsoleLabel',
-          handler: () => {}
-        }
-      ]}
-    />
-  )
-}
-
-const VmActionBar = ({
-  vm,
-  handlers
-}) => {
-  const ActionBar = vmActionBarByState[vm.power_state]
-
-  if (!ActionBar) {
-    return <p>No action bar for state {vm.power_state}</p>
-  }
-
-  return <ActionBar vm={vm} handlers={handlers} />
-}
+import VmActionBar from './action-bar'
 
 // ===================================================================
